standalone_connectlife/requirements.txt
aiohttp>=3.8.0
pytz>=2023.3
python-dotenv>=1.0.0
